requirements.txt
streamlit==1.28.0
PyPDF2==3.0.1
python-docx==1.1.0
pymupdf==1.23.8
groq==0.4.0
pinecone-client==3.0.1
sentence-transformers==2.2.2
pandas==2.1.4
fpdf2==2.7.7
torch==2.1.0
transformers==4.35.2
numpy==1.24.3
pillow==10.1.0
requests==2.31.0
urllib3==1.26.18